services/content-pipeline/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
databases>=0.7.0
aiosqlite>=0.19.0
sqlalchemy>=2.0.0

# API clients
httpx>=0.24.0
openai>=1.0.0
anthropic>=0.18.0

# Pipeline orchestration（worker 用 LangGraph 编排 6 阶段，替代原 for-loop）
langgraph>=0.2.0

# Content processing
python-multipart>=0.0.6
Pillow>=10.0.0
beautifulsoup4>=4.12.0
markdown>=3.4.0

# Browser automation
playwright>=1.40.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0